# lib/api-zod/src/generated/types/leaderboardPage.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CodeMaster coding-practice platform API
 * OpenAPI spec version: 0.1.0
 */
import type { LeaderboardEntry } from './leaderboardEntry';

export interface LeaderboardPage {
  items: LeaderboardEntry[];
  page: number;
  pages: number;
  total: number;
}