# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CodeMaster coding-practice platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileRole } from './profileRole';
import type { SubmissionSummary } from './submissionSummary';

export interface Profile {
  id: number;
  username: string;
  name: string;
  /** @nullable */
  avatar: string | null;
  role: ProfileRole;
  solvedCount: number;
  totalSubmissions: number;
  acceptedSubmissions: number;
  acceptanceRate: number;
  currentStreak: number;
  joinedAt: Date;
  recentSubmissions: SubmissionSummary[];
}